+++ CHANGELOG.md
@@ -275,5 +275,6 @@
 * Define `interruptable` method for `Continuation`
 * Define `Catchable` typeclass to catch errors from `Conclusion`
 * Define `repeat` method of `Stream` datatype
-* Define `subset` and `cartesian` methods of `Set` interface
+* Define `subset` method of `Set` interface
 * Define `satisfy` method in `Predicate` module
+* Make `fold` and `find` stateful expressions